#1 Feat : 테이스팅 노트 작성 후 와인 평균값 노트 업데이트 및 조회 기능 구현

+++ src/main/java/com/drinkeg/drinkeg/domain/TastingNote.java
@@ -33,7 +33,7 @@ public class TastingNote {
 
     private String color;
 
-    // 점수 1 ~ 5
+    // 점수 0 ~ 5
     private int sugarContent;
     private int acidity;
     private int tannin;
@@ -50,7 +50,8 @@ public class TastingNote {
     @Convert(converter = StringListConverter.class)
     private List<String> scentFinish = new ArrayList<>();
 
-    private int satisfaction;
+    // 만족도 0 ~ 5, 소수점 가능
+    private float satisfaction;
 
     private String memo;
 }
